Extract components from Home page
