+++ src/ensembl/src/content/home/components/homepage-site-info/HomepageSiteInfo.scss
@@ -1,0 +1,23 @@
+@import 'src/styles/common';
+
+.siteMessage {
+  margin-left: 120px;
+  padding-top: 2%;
+
+  h4 {
+    color: $grey;
+    font-size: 13px;
+    font-weight: $normal;
+    margin-bottom: 0.8em;
+  }
+
+  p {
+    font-size: 13px;
+    margin: 0;
+    line-height: 1.5;
+  }
+
+  .convoMessage {
+    margin-top: 1em;
+  }
+}
