+++ src/ensembl/src/content/home/components/homepage-species-bar/HomepageSpeciesBar.scss
@@ -1,0 +1,8 @@
+.emptySpeciesBar {
+  display: flex;
+  flex-direction: column;
+
+  .speciesSelectorBannerLink {
+    font-weight: 700;
+  }
+}
